fix(kube-control-plane): use patches for node specific configuration
Move setting the `bind-address` to the Node's IP  configuration from the
ClusterConfiguration object to kubeadm patches.

The kubeadm ClusterConfiguration is used for configuration common to all
the nodes, while specific node configuration must me done using patches,
as per kubeadm documentation.

Since kubeadm 1.31, kubeadm does not use ClusterConfiguration for the
upgrades and uses the configuration stored in the ConfigMap instead.
With patches we make sure the node-specific config is always applied.

Patches must be appliead both at init-time and upgrade-time.